gui-manager/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.35
ptyprocess==0.7.0; sys_platform != "win32"
pywinpty>=2.0.0; sys_platform == "win32"
pyyaml==6.0.2
python-dotenv==1.0.1
python-multipart==0.0.12
aiofiles==24.1.0
websockets==13.1
